third_party/blink/web_tests/external/wpt/css/css-properties-values-api/registered-property-computation-expected.txt

This is a testharness.js-based test.
Found 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] <color> values are computed correctly [color-mix(in srgb, black, white)]
  assert_equals: expected "color(srgb 0.5 0.5 0.5)" but got "color-mix(in srgb, black, white)"
[FAIL] <color> values are computed correctly [color-mix(in srgb, currentcolor, red)]
  assert_equals: expected "color-mix(in srgb, currentcolor, color(srgb 1 0 0))" but got "color-mix(in srgb, currentcolor, red)"
Harness: the test ran to completion.
